--- a/build.sbt
+++ b/build.sbt
@@ -92,7 +92,7 @@ lazy val stdSettings: Seq[sbt.Def.SettingsDefinition] = Seq(
     "dev.zio"                %% "zio-streams"                 % zioVersion,
     "dev.zio"                %% "zio-test"                    % zioVersion % "test",
     "dev.zio"                %% "zio-test-sbt"                % zioVersion % "test",
-    "dev.zio"                %% "zio-interop-reactivestreams" % "1.3.3",
+    "dev.zio"                %% "zio-interop-reactivestreams" % "1.3.4",
     "dev.zio"                %% "zio-logging"                 % "0.5.6",
     "ch.qos.logback"          % "logback-classic"             % "1.2.3",
     "org.scala-lang.modules" %% "scala-collection-compat"     % "2.4.3",
@@ -117,7 +117,7 @@ lazy val interopFutures = (project in file("interop-futures"))
     name := "zio-kinesis-future",
     assemblyJarName in assembly := "zio-kinesis-future" + version.value + ".jar",
     libraryDependencies ++= Seq(
-      "dev.zio" %% "zio-interop-reactivestreams" % "1.3.3"
+      "dev.zio" %% "zio-interop-reactivestreams" % "1.3.4"
     )
   )
   .dependsOn(core)