Chunk has more efficient append performance than prepending to a List and reversing and integrates better with ZIO streams.

A rough comparison (`ProducerTest` - "efficiently") shows about 20% higher throughput and 30% less CPU usage.

```
Without Chunks:

interval=298 s, total records published=7219000, throughput=24187.33373092722 records/s, success rate=100,00 %, failed attempts=0, shard prediction errors=0, mean latency=7077 ms, 95% latency=8495 ms, min latency=379 ms, 2nd attempts=0, max attempts=1, mean latency=7077 ms, mean payload size=4355 bytes, mean record size=2473 bytes, nr PutRecords calls=23279, mean nr PutRecords calls=77.99652887134711 calls/s

With Chunks:

interval=282 s, total records published=8000000, throughput=28278.643614859015 records/s, success rate=100,00 %, failed attempts=0, shard prediction errors=0, mean latency=6432 ms, 95% latency=8059 ms, min latency=333 ms, 2nd attempts=0, max attempts=1, mean latency=6432 ms, mean payload size=3878 bytes, mean record size=2517 bytes, nr PutRecords calls=28876, mean nr PutRecords calls=102.0717641278336 calls/s
```